% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{a logical vector of predicted values.}

\item{truth}{a logical vector of true values.}
}
\value{
a numerical value for accuracy
}
\description{
Accuracy is defined as (# TP + # TN) / (# All Positives + # All Negatives).
}
\examples{
truth = c(T,T,T,T,T,F,F,F,F,F)
pred = c(T,F,T,T,F,T,T,T,F,F)
accuracy(pred,truth)
}
