% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{Adjusted R-squared}
\usage{
adjR2(pred, truth, d = 1)
}
\arguments{
\item{pred}{a (non-empty) numeric vector of predicted values of the dependent variable in regression.}

\item{truth}{a (non-empty) numeric vector of true values of the dependent variable in regression.}

\item{d}{number of independent variables in the regression model. Default value is 1}
}
\value{

}
\description{
Adjusted R-squared is the fraction of variance explained by the model, similar to R-suared. It also penalizes for the number of independent variables in the regression model.
}
\examples{

}
