% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSquareCounts.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{chi-square test for count data}
\usage{
chiSquareCounts(tib)
}
\arguments{
\item{tib}{a dataframe or tibble with 2 columns of interest}
}
\value{
a matrix of counts and p-value of the chi-square test
}
\description{
Implement a chi-squared test that takes a dataset with 2 columns of categorical variable to return the counts matrix and p-value for the chi square test.
}
\examples{
data = tibble::read_csv('https://jlucasmckay.bmi.emory.edu/global/bmi585/demographics.csv')
data = subset(data, select = c('sex','group'))
chi_sq(data)
}
