% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectSize.R
\name{effectSize}
\alias{effectSize}
\title{Effect size}
\usage{
effectSize(x, g)
}
\arguments{
\item{x}{An atomic vector of numeric data}

\item{g}{An atomic vector of strings or factors (categorical data)}
}
\value{
a numeric value for effect-size or Cohen's d
}
\description{
One very useful measure of effect size is Cohen’s d. This is the difference in means between two groups divided by some measure of the standard deviation. Effect sizes are one of the most common parameters used in power analysis.
}
\examples{
set.seed(24)
x = c(rnorm(50,mean=10,sd=2),rnorm(50,mean=3,sd=6))
g = c(rep('M',50),rep('F',50))
effectSize(x,g)
}
