% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{F1 score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{a logical vector of predicted values.}

\item{truth}{a logical vector of true values.}
}
\value{
a numerical value for accuracy
}
\description{
F1 score is defined as 2 x (precision x recall) / (precision + recall).
}
\examples{
truth = c(T,T,T,T,T,F,F,F,F,F)
pred = c(T,F,T,T,F,T,T,T,F,F)
f1(pred,truth)
}
