% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrAdjust.R
\name{fdrAdjust}
\alias{fdrAdjust}
\title{FDR: multiple comparison adjustment}
\usage{
fdrAdjust(p, alpha = 0.05)
}
\arguments{
\item{p}{a (non-empty) numeric vector of p-values.}

\item{alpha}{significance level}
}
\value{
a logical vector of adjusted p-values < alpha.
}
\description{
Implementation of FDR function for multiple hypothesis testing
}
\examples{
pval <- c(0.0050,0.0025,0.0075,0.0100,0.0125,0.0150,0.0175,0.0200,0.0225,0.0250)
fdrAdjust(pval) # TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
}
