% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumN.R
\name{minimumN}
\alias{minimumN}
\title{Minimum sample size mapped from effect size (Cohen's d)}
\usage{
minimumN(d, power = 0.8)
}
\arguments{
\item{d}{true difference in means}

\item{power}{power of test (1 minus Type II error probability), default 0.8}
}
\value{
number of observations (per group)
}
\description{
A wrapper function around power.t.test that maps effect size (Cohen's d) to minimum sample size with the default power of 0.8.
}
\details{
User's can change the power from 0.8.
}
\examples{
minimumN(d = 1)
}
