% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcApprox.R
\name{pcApprox}
\alias{pcApprox}
\title{PCA Approximation}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{a dataframe or tibble of numerical data.}

\item{npc}{number of principal compnents}
}
\value{
tibble; approximation to the data \code{x} based on \code{npc} PCs
}
\description{
A function that takes a numerical object from scale and reverses the centering/scaling
}
\examples{
data = iris[,1:4]
pcApprox(data,3)
}
