% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcLollipop.R
\name{pcLollipop}
\alias{pcLollipop}
\title{PCA Lollipop plot}
\usage{
pcLollipop(x)
}
\arguments{
\item{x}{a dataframe or tibble of numerical data.}
}
\value{
a plot object with lollipop plot as output
}
\description{
Lollipop plot is for visualization of Principal Component loadings
}
\examples{
data = iris[,1:4]
pcApprox(data,3)
}
