% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postHocPower.R
\name{postHocPower}
\alias{postHocPower}
\title{Post-hoc power}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d}{a numeric value for Cohen's d (effect size)}

\item{n1}{an integer for sample size of group 1}

\item{n2}{an integer for sample size of group 2}
}
\value{

}
\description{
Estimation of power based on 1000 simulations from  a measure of Cohen’s d and an n per group.
}
\examples{
set.seed(2021)
postHocPower(0.2,1000,1500)    # 0.974
}
