% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{R-squared}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{a (non-empty) numeric vector of predicted values of the dependent variable in regression.}

\item{truth}{a (non-empty) numeric vector of true values of the dependent variable in regression.}
}
\value{

}
\description{
R-squared is the fraction of variance explained by the model. It gives you an idea of how many data points fall within the line of the regression equation.
}
\examples{
model = lm(Sepal.Width ~ Sepal.Length+Petal.Length, data = iris)
truth = iris$Sepal.Width
pred = model$fitted.values
r2(pred,truth)
}
