% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivity}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{a logical vector of predicted values.}

\item{truth}{a logical vector of true values.}
}
\value{
a numerical value for sensitivity
}
\description{
Sensitivity(pred,truth) : Sensitivity is defined as # True Positives / # All Positives = # True Positives / (# True Positives + # False Negatives). Sensitivity is also referred to as recall.
}
\examples{
truth = c(T,T,T,T,T,F,F,F,F,F)
pred = c(T,F,T,T,F,T,T,T,F,F)
sensitivity(pred,truth)
}
