% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{a logical vector of predicted values.}

\item{truth}{a logical vector of true values.}
}
\value{
a numerical value for specificity
}
\description{
Specificity is defined as # True Negatives / # All Negatives.
}
\examples{
truth = c(T,T,T,T,T,F,F,F,F,F)
pred = c(T,F,T,T,F,T,T,T,F,F)
specificity(pred,truth)
}
