% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSidedT.R
\name{twoSidedT}
\alias{twoSidedT}
\title{Area under the t-density}
\usage{
twoSidedT(t, n)
}
\arguments{
\item{t}{An integer or numeric value for the test-statistic}

\item{n}{An integer value for degrees of freedom}
}
\value{
a p-value or area under the t-statistic
}
\description{
Convert a test statistic into the area under the t-densities or P value
}
\examples{
twoSidedT(-.785,14)
}
