% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSidedZ.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{Convert a test statistic into the area under the z-density}
\usage{
twoSidedZ(z)
}
\arguments{
\item{z}{An integer or numeric value for the test-statistic}
}
\value{
a p-value or area under the z-statistic
}
\description{
Convert a test statistic into the area under the z-densities or the P value
}
\examples{
twoSidedZ(-.785)
}
