% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unscale.R
\name{unscale}
\alias{unscale}
\title{Unscale}
\usage{
unscale(x)
}
\arguments{
\item{x}{A scaled object with scale.default, the centered, scaled matrix and the numeric centering and scalings (stored as attributes "scaled:center" and "scaled:scale")}
}
\value{
An unscaled numeric vector or matrix
}
\description{
A function that takes a numerical object from scale and reverses the centering/scaling
}
\examples{
set.seed(1)
x = matrix(sample(1:12), ncol= 3)
xs = scale(x, center = TRUE, scale = TRUE)
unscale(xs)
}
