% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welchT.R
\name{welchT}
\alias{welchT}
\title{Welch's t-test}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}
}
\value{
a numeric vector with p-value, estimated t-statistic and degrees of freedom
}
\description{
Performs Welch's t-test for two group dataset
}
\examples{
welchT(1:10, y = c(7:20)) # 0.000019
welchT(1:10, y = c(7:20, 200)) # 0.124513
}
